CHG: Moved newton method to root_finding.py, addressed wrong e constants, updated constants, and changed x and r into one augmented state in encke motion
