+++ constants.py
@@ -2,9 +2,6 @@
 constants.py contains all the commonly used constants throughout the software
 """
 
-import numpy as np
-
-G = 6.6743 * 10**-11 # m^3 / (kg * s^2)  --- Gravitational Constant
-M = 5.97219 * 10**24 # kg   ---   Earth's Mass
-mu = G*M # m^3 / s^2   ---   G*M,the gravitational parameter
-e = np.e
+G_m3_kgs2 = 6.6743 * 10**-11 # m^3 / (kg * s^2)  --- Gravitational Constant
+M_kg = 5.97219 * 10**24 # kg   ---   Earth's Mass
+EARTH_MU_m3_s2 = G_m3_kgs2*M_kg # m^3 / s^2   ---   G*M,the gravitational parameter
